added load promoFilm
