+++ project/src/components/app/app.tsx
@@ -1,6 +1,5 @@
 import {Routes, Route} from 'react-router-dom';
 import {AppRoute} from '../../const';
-import { checkAuthAction } from '../../store/api-action';
 import PrivateRoute from '../../components/private-route/private-route';
 import SingIn from '../../pages/sing-in/sing-in';
 import Main from '../../pages/main/main';
@@ -11,13 +10,10 @@ import Player from '../../pages/player/player';
 import Error from '../error/error';
 import Loader from '../loader/loader';
 import { BrowserRouter } from 'react-router-dom';
-import { useAppSelector, useAppDispatch } from '../../hooks';
+import { useAppSelector } from '../../hooks';
 
 
 function App() {
-  const dispatch = useAppDispatch();
-  dispatch(checkAuthAction());
-
   const isDataLoaded = useAppSelector((state) => state.isDataLoaded);
 
   if (isDataLoaded) {
