+++ project/src/index.tsx
@@ -3,12 +3,13 @@ import { Provider } from 'react-redux';
 import ReactDOM from 'react-dom';
 import App from './components/app/app';
 import {store} from './store/index';
-import {loadPromoFilmAction, loadFilmsAction} from './store/api-action';
+import {loadPromoFilmAction, loadFilmsAction, checkAuthAction} from './store/api-action';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
 store.dispatch(loadPromoFilmAction());
 store.dispatch(loadFilmsAction());
+store.dispatch(checkAuthAction());
 
 ReactDOM.render(
   <React.StrictMode>
