+++ project/src/pages/add-review/add-review.tsx
@@ -1,6 +1,7 @@
 import { Link } from 'react-router-dom';
 import Logo from '../../components/logo/logo';
 import Review from '../../components/review/review';
+import User from '../../components/user/user';
 import { AppRoute } from '../../const';
 import { useAppSelector } from '../../hooks';
 
@@ -40,16 +41,7 @@ function AddReview({filmId = 0}: typeProps) {
             </ul>
           </nav>
 
-          <ul className="user-block">
-            <li className="user-block__item">
-              <div className="user-block__avatar">
-                <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-              </div>
-            </li>
-            <li className="user-block__item">
-              <Link to={AppRoute.Login} className="user-block__link">Sign out</Link>
-            </li>
-          </ul>
+          <User />
         </header>
 
         <div className="film-card__poster film-card__poster--small">
