+++ project/src/pages/film/film.tsx
@@ -4,6 +4,7 @@ import Logo from '../../components/logo/logo';
 import CardList from '../../components/card-list/card-list';
 import {Link, useLocation} from 'react-router-dom';
 import Tabs from '../../components/tabs/tabs';
+import User from '../../components/user/user';
 import { useAppSelector } from '../../hooks';
 
 type typeProps = {
@@ -38,16 +39,7 @@ function Film({filmId = 0}: typeProps) {
           <header className="page-header film-card__head">
             <Logo />
 
-            <ul className="user-block">
-              <li className="user-block__item">
-                <div className="user-block__avatar">
-                  <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-                </div>
-              </li>
-              <li className="user-block__item">
-                <Link to={AppRoute.Login} className="user-block__link">Sign out</Link>
-              </li>
-            </ul>
+            <User />
           </header>
 
           <div className="film-card__wrap">
